@@ -3,7 +3,7 @@ description: Internal package to create platform-appropriate HTTP clients.
 version: 1.0.0
 
 environment:
-  sdk: ^3.0.0
+  sdk: ^3.2.0
 
 dependencies:
   http: ^1.1.0
