@@ -3,17 +3,17 @@ description: Internal data package of WebTrit Phone application.
 version: 1.0.0
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
 
 dependencies:
   clock: ^1.1.1
-  drift: ^2.13.0
+  drift: ^2.13.1
   io: ^1.0.4
   path: ^1.8.2
-  sqlite3_flutter_libs: ^0.5.17
+  sqlite3_flutter_libs: ^0.5.18
 
 dev_dependencies:
   build_runner: ^2.4.6
-  drift_dev: ^2.13.1
+  drift_dev: ^2.13.2
   lints: ^3.0.0
   test: ^1.24.9
