@@ -3,8 +3,8 @@ description: Library to extract/get the application info from Apple App Store an
 version: 1.0.0
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
-  flutter: ^3.10.0
+  sdk: '>=3.2.0 <4.0.0'
+  flutter: ^3.16.0
 
 dependencies:
   flutter:
