@@ -4,20 +4,20 @@ version: 0.0.1
 homepage:
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
-  flutter: ^3.13.0
+  sdk: ">=3.2.0 <4.0.0"
+  flutter: ^3.16.0
 
 dependencies:
   flutter:
     sdk: flutter
 
   equatable: ^2.0.5
-  uuid: ^3.0.7
+  uuid: ^4.2.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.3
+  flutter_lints: ^3.0.1
 
   pigeon: ^11.0.1
 
