@@ -3,8 +3,8 @@ description: Flutter package for processing phone numbers.
 version: 0.0.1
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
-  flutter: ^3.10.0
+  sdk: '>=3.2.0 <4.0.0'
+  flutter: ^3.16.0
 
 dependencies:
   flutter:
@@ -14,4 +14,4 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^3.0.0
+  flutter_lints: ^3.0.1
