@@ -4,7 +4,7 @@ publish_to: none
 version: 1.0.0
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
 
 dependencies:
   equatable: ^2.0.5
