@@ -38,10 +38,10 @@ class LoginCoreUrlAssignTab extends StatelessWidget {
         }
       },
       builder: (context, state) {
-        return WillPopScope(
-          onWillPop: () async {
+        return PopScope(
+          canPop: false,
+          onPopInvoked: (didPop) {
             _onCoreUrlAssignBack(context);
-            return false;
           },
           child: LoginScaffold(
             appBar: AppBar(
